Convert LIFT IDs to NFC when storing them in Mongo
FDO objects are NFD internally, but we need to store LIFT IDs in a form
that matches what they will be on disk. That will almost certainly be
NFC, since most LIFT files that we deal with will have been exported
from FDO. If we get any LIFT files from other sources, it's possible
that there could be problems, but we'll deal with those later.