Wrapper script now finds and kills hung processes
If a previous LfMerge run hung up on the "About to dispose FW project
(projectcode)" step, the lfmergeqm script will now detect and kill the
hung process, and remove the lockfile. This will result in the hung
process being run a second time (since it still has an entry in the sync
queue), and the second run is 99.9% likely to work.

This does not address the root cause of the issue, but it treats the
symptoms quite well.