Strip br tags when found in FDO data
Since the text <br/> and <br /> should not be found in any genuine FW
data, it's safe to assume that this text was inserted by our LfMerge
bug. So we'll strip it from incoming FW data, to fix the incorrect data
that the bug introduced earlier.

This commit also contains unit tests to ensure that the <br/> stripping
is handled correctly and doesn't contaminate any other data.