Use pinned FlexBridge build to avoid test failures
A recent change to FlexBridge includes using ###.### format for branch
numbers in Mercurial (e.g., 7500002.7000072 instead of 7000072). This is
causing some test failures in our SynchronizeAction tests, which don't
handle this change correctly. Until the tests can be brought up-to-date
with the latest changes in FlexBridge, we can just use a pinned version
from before that change was introduced.