Don't call string.Format twice
The logger code already calls string.Format on the message and
parameters it receives. This is sometimes causing an error about "Input
string was not in a correct format" in the second call, when the first
string.Format call happens to produce output that contains formatting
characters that the second string.Format thinks it should handle. The
result is swallowed exceptions. By removing one call to string.Format,
we should be able to see the actual exceptions.