Correct example sentence magic string
When reviewing code on a different PR, I noticed that the value of LFFieldNameForExampleSentence was set to definition.  This looks wrong and is likely the result of a copy/paste error (from the line above).  I have no way to test this, and this PR doesn't fix a known bug.  However, it seems so obviously wrong that I got up the guts to create this PR :)