Debian packaging scripts from ci-builder-scripts
This contains the Debian packaging scripts from our ci-builder-scripts
repo, so that we can build packages the same way the Jenkins build
server does. I copied the scripts as they were a week ago, just before a
commit that removed Xenial from the list of build targets, because we may
still want to target Xenial with our LfMerge build, at least for a while.