WIP: Improve Debian build
Working on getting the Debian build to actually produce packages. Other
possible improvements would be to create a baseline Docker image for a
package builder with sbuild, to save a lot of time on the setup of the
initial sbuild tarballs that download a lot of the base Debian packages.
A Docker image that contains those tarballs already ready to go would
help a lot, by allowing the base sbuild image to be pre-built and ready
to go instantly.