Use dotnet-gitversion tool in FW8 build (#113)
At some point recently, GitVersion changed the name of the `master`
branch to `main`, and if we run `dotnet gitversion` with our current
config we get a KeyNotFoundException with the message "The given key
'master' was not present in the dictionary." Changing `master` to
`main` everywhere except the branch-name-matching regex fixes this.

This commit also switches to using mono5-sil since mono-sil is no
longer able to build (its moz-roots tool, used in the build, tries
to download certificate data from a URL that no longer exists).

Finally, we pin a FLExBridge build in TeamCity so that we can download
FLExBridge DLLs from a build that's known to work, even if TeamCity
marked the build as a failure. (The last successful build in TeamCity
did not produce correct DLLs). This pin will be removed in the future.