WIP: Debian packaging worked for the first time
Using the dbuild.sh script in this commit, I was able to build a Debian
package with a 0.1.0 version number. Next step is to start bringing the
build more in line with the Jenkins build steps, such as getting the
version number of the Debian package from GitVersion. But this worked,
so I'm making a commit so that I can always revert back to what worked.