WIP: No need to build Mercurial in Docker
We can download the Mercurial-x86_64.zip file from TeamCity and that
works. It also has the contents that the debian/rules.in file expects
(i.e., no other directories inside the "mercurial" dir) so there's no
need to patch debian/rules.in if we omit this build step.

So this reduces complexity while still building correctly.