WIP: Use DOCKER_BUILDKIT to save SSD churn
Run unit tests in a tmpfs to save SSD churn from all the temporary files
being created and deleted in unit tests. Doesn't save much time if you
have a fast SSD, but does save SSD lifespan. Requires Docker 18.09 or
above, and that the DOCKER_BUILDKIT environment variable is set to 1.
But since BuildKit is available in GitHub Actions, this isn't a major
limitation as it won't restrict where this can run.