Base build image still runs as root
Turns out that something in the mcr.microsoft.com/dotnet/sdk:5.0 image
is causing sudo to not have effective uid 0, which makes it fail. Rather
than tracking that down, we'll just remain as root for a little bit
longer in the Docker setup, and switch to the builder user later.