Fix package build
During a package build when we call CompileOnly we have to restore
packages even when they already exist in `$(RootDir)/packages`. The
reason is that the source package doesn't include the `obj`
directories which contain some information about the packages that
the build needs to find dependencies. However, those files in `obj`
contain absolute paths so we can't add them to the source package.
Instead we call RestorePackages again which will recreate the `obj`
directories, but doesn't download anything from the Internet.

Also, the working directory seems to be different so we have to pass
in a full path as the location for the packages.