Don't copy Mercurial docs into package
Depending on where we get the Mercurial_x64.zip file from, it may or may
not have the Mercurial/doc folder, and if that folder is missing, the
build fails. And since the Mercurial docs aren't serving a useful
purpose in the LfMerge package, it's best to just leave them out.