Send/Receive can now create canonical LF Tags list
FLEx projects will now have a custom list created if it doesn't already
exist, using a GUID reserved for this purpose, for storing Language
Forge tags. We should probably put something in the list description
like "Please do not edit this list", but since it will be re-created if
it doesn't exist, that's probably not needed.