No more checking out the Dockerfile in FW8 builds (#153)
The fieldworks8-master branch now needs its own version of the
Dockerfile, because it needs to have some code (adjusting the
AssemblyInfo.cs files) that's not needed in the master branch. Checking
out the master branch copy was causing PR 152 not to actually build what
I thought it was building.