Update AssemblyInfo files manually in FW8 builds
The GenerateAssemblyInfo feature came in with .Net Core 3.1. Anything
built with Mono 5 won't have it, so for the FW8 builds we need to update
AssemblyInfo.cs files ourselves. I haven't found a good GHA action for
that, so we'll do it ourselves with a simple bash script. The bash
script will reorder lines in the file, but that's not a big deal, and it
keeps it very simple.