Create GitHub release automatically (#158)
The first release will use a manually-created body. After that I will
replace the `body: ...` in the YAML with `generate_release_notes: true`
which will use GitHub's handy auto-generated release notes, that consist
of one line per PR with the PR name and author (and a link to the PR),
plus a link to the differences between the previous tag and this one.
But I want at least one release with a manually-generated content.