Sync fieldworks8-master and master workflows (#159)
The "should this workflow run" conditional has had a minor change in the
fieldworks8-master branch that was not yet pulled into the master
branch. While we're at it, we shouldn't dump the github context to the
logs carelessly, because it does contain some sensitive info like the
GITHUB_TOKEN value. It's automatically redacted in the GitHub logs, but
in case a bug makes that redaction fail, we don't want to expose it.