Make workflow safe to run locally (no git reset)
If running the workflow locally using `act` or `docker run`, we don't
want any `git reset --hard HEAD` commands being run on the mounted repo,
as that could easily blow away uncommitted work in the developer's own
copy of the repo. So we mount the repo somewhere else, and then copy the
repo into a working location that we can safely run `git reset` inside.