Local build now figures out FW8 and FW9 branches
This isn't needed in the GitHub Actions workflow since that runs in
separate containers with their own individually-checked-out copies of
the repo, but in the local build we only have one checked-out branch,
and we need to figure out the other branch to build. E.g. if the
checked-out branch comes from a FW9 branch, then we need to calculate
which FW8 branch to build (and default to fieldworks8-master). If the
checked-out branch is a FW8 branch, then we need to figure out which FW9
branch to build, and default to master.