Remove deleted files from .csproj list
This wasn't needed in the master branch version of this change, where it
uses the modern .NET build system that doesn't need a list of every
file. But here in the fieldworks8-master branch, the .csproj file
includes a list of every file in the project, so when we delete a file
we also need to remove it from the .csproj list as well.