No more calling PHP code from LfMerge (#173)
This PR finally implements the updateCustomFieldViews logic in LfMerge,
as part of the SetCustomFieldConfig method in MongoConnection (which is
where the logic belongs anyway). This will allow us to move the lfmerge
package into its own container in Language Forge, instead of having it
so tightly coupled to the PHP container.