Speed up build by using gzip instead of xz (#177)
Since we're not going to upload the source packages the build process
creates, there's no sense in sacrificing time for better compression
ratios. Instead, we'll use gzip at the default compression value (6),
which is a lot faster and produces files that are 40% larger than what
xz produces. We'll continue to use xz on the binary package files; it's
just the orig.tar.xz (now orig.tar.gz) files where we want to use lower
compression to save time because they're so large.