Start converting debian build into Docker image
This commit takes the Debian rules.in file, and starts the process of
turning it into a shell script that will install the appropriate files
into an empty directory which will be turned into a tarball to use in a
Docker container layer.

In this commit, we've deliberately kept the tabs and most other syntax
from the rules.in file intact, even though it's not valid Bash syntax,
so that most lines will be similar and Git will detect the rename.