Remove caching for final lfmerge image build
For some unknown reason, caching is failing in this step even though it
succeeds in the earlier steps. The reason given is:

buildx failed with: error: cache export feature is currently not supported for docker driver. Please switch to a different driver (eg. "docker buildx create --use")

I have no idea why this step would be running a different driver than
the previous steps, as the configuration is the same AFAIK. But we'll
remove the caching lines and see what happens.