Do not cause release to fail if tag already exists (#216)
There are some circumstances where the release tag being pushed already
exists; most notably, a bugfix that only applies to FW 9 or FW 8 and
therefore was only applied to one branch. The other branch would build
the same tag over again, and should not fail just because the tag can't
be created a second time.