Fix LfMerge settings from env vars
Turns out `sudo` resets environment variables except for those appearing
on a whitelist. So we ensure that all LFMERGE_* variables appear on the
whitelist in the LfMerge container.

We also move the SetAllMembers() call in LfMergeSettings into the
constructor, because there's no reason for it to be in Initialize().

These two changes are enough to fix the bugs that were causing LfMerge
settings not to be picked up by the process.