Slimmer LfMerge images (#236)
We now build from a base image stored in the GitHub Container Registry.
By storing the base image in GHCR and only updating it when once of the
lfmerge dependencies changes (which almost never happens), we can avoid
re-running `apt-get update` all the time, which always ends up causing a
layer change and the entire 700MB layer to be different. This results in
much slimmer lfmerge images and faster download times.