Cleaning up old TODOs: keep string.Empty here
Although set_String would indeed function if passed a null, at least in
its current state, there's no guarantee that it wouldn't be changed at
some point, and the null-check and replacing with string.Empty costs
practically nothing. So we'll keep it for safety, and get rid of the
TODO comment that implies it might change at some point.