Cleaning up old TODOs: Decided against this one
Decided that the method recommended in that Stack Overflow answer (a
generic FactoryConverter class derived from JsonConverter) is way too
much abstraction, and the simple approach here is much better. More
readable, and less mysterious. If we were needing to do this hundreds of
times we'd have to take the FactoryConverter approach, but we don't.