Wait for queue file to be closed, not opened (#257)
Running the queue manager when a file is created in the directory is
sometimes triggering too early, because we're writing to these files via
`echo "projectCode: something" | sudo tee syncqueue/something`. But
programs like `tee` have a three-step process for writing to a file:

1) Open the file
2) Wait for contents from stdin, then write them out
3) When stdin closes, close the file

This means that by simply listening for the `create` event, we're
sometimes triggering lfmergeqm to run between steps 1 and 2, and then
the queue manager doesn't find any contents in the file in the queue
directory, and exits again -- and then that project doesn't run the
Send/Receive process until a different project triggers Send/Receive.

By waiting for the close_write event instead, we'll ensure that any
files we are notified about will already have their contents.