Exit build if any errors occur (#264)
We'll use `set -e` in all our build scripts, to make sure we had no
errors. That's an omission we should have rectified long ago.

Note that some people recommend `trap some_func ERR` instead, to call
`some_func` if any errors occur, which would allow recovery in some
circumstances. I'll use that it if turns out to be useful, but exiting
on error is the most helpful behavior most of the time in this build.