Update dotnet SDK and runtime images to 6.0 (#266)
Since we're changing the base image anyway, this is a perfect time to
upgrade to .NET 6 as it's a Long-Term Support release.

This will mean a longer build and a longer deployment the first time,
because the 700-megabyte layer in the base images will be new. But now
that the base images are updated to .NET 6, I won't need to do that
again anytime soon. And since we'd have to download new base images
anyway now that we differentiate between the SDK and runtime images,
this is the perfect time to make that switch.