Replacing usage of mono with dotnet and using its nuget
Couple notes:
Removed RestoreBuildTasks. It doesn't appear to be necessary -- Nuget and the relevant projects should handle this if it turns out it is.
Didn't replace --debug flag in mono with anything. It's unclear if the theoretical dotnet equivalent would perform as desired, but we could maybe use "verbosity" if needed.

These changes are experimental, but a commit is required to test, since in docker creation pending changes are cleared.