Try to copy Mono.Posix.DLL into output folder
With the .NET 6 build, the Mono.Posix DLL is ending up in a different
location than the other DLLs, along with a native library (in .so format
for Linux) that it probably also needs, so we'll add those files to the
installation script.

Still doesn't work, but it was worth a try.