Delete TODO comment that will not be done
Decided not to implement this idea, as there are too many possibilities
where that heuristic is not useful. E.g., only one entry was changed,
so the for loop skips all other entries for not having an updated date;
if that one entry throws an exception, should we treat this as 100% of
the entries being bad? No.