Log to stdout and stderr by default (#134)
Can still log to syslog, if desired, by setting environment variable
LFMERGE_LOGGING_DEST to "syslog". LFMERGE_LOGGING_DESTINATION is also
accepted as an env var name. Any other value than "syslog", including
the environment variable being entirely absent, means log to console.

In the future LFMERGE_LOGGING_DEST may be used to specify the name of a
file that should be used to log LfMerge output. If that is implemented,
then a value of "-" (or an absent env var) will be the only one that
means "log to console", so it's best to either omit the env var entirely
or else explicitly set it to "-".

The LFMERGE_LOGGING_STDERR_THRESHHOLD environment variable can be used
to configure the severity threshhold at which log messages are sent to
stderr instead of stdout. Default is Warning, but Error and Notice are
also sensible values. Values are case insensitive, and numbers can be
used as well (Error = 3, Warning = 4, Notice = 5).