Mount user's home nuget folder as cache
Instead of using a non-standard `/storage/nuget` folder that won't be on
most people's computers, use the NuGet package cache that they alreday
have, located in `~/.nuget/packages`. Because if they're using VS Code
or any other IDE that knows about .NET and NuGet, that folder is
probably already populated with the appropriate packages, which makes
even the first build go faster.