VS Code launch settings for debugging unit tests
To use this, choose the "Run Test Task" command in VS Code. The dotnet
test will run, then pause waiting for a debugger to attach. Go to the
debug menu and choose the "Attach" configuration, then enter the PID
printed on the test console. (There will be two, one for LfMerge.Tests
and one for LfMerge.Core.Tests; most of the tests are in LfMerge.Core so
that's the one you'll need most of the time). The debugger will attach
to the test process, which is still in a paused state. Set any
breakpoints you want to hit in the unit tests, then unpause the debugger
and it will start running the unit tests.