Stop caching in LfMergeSettings so unit tests work
LfMergeSettings was caching the value of environment variables, but some
strings like BaseDir were being accessed in the constructor, before
derived classes like LfMergeSettingsDouble had a chance to change them.
This meant that many tests, which relied on testlangproj or other
projects being copied into the test's temp dir, were failing because
LfMergeSettings was using the default /var/lib/languageforge/... base
directory instead. By removing the caching, we ensure that the
LfMergeSettingsDouble class is able to change the BaseDir correctly.