Allow testing uncommitted code
By removing the `git checkout` and `git clean` steps, we ensure that we
can run the build process against uncomitted code in the repository.
This will make testing changes much simpler.

A leter commit will probably remove the copying of the Git repo into a
temporary working directory, and simply run the build in the user's repo
via a bind mount. That will mean that build outputs (such as TestResult
directories) will end up in your repo, which will be useful in any
situation where you want to debug the unit tests.