Remove entire calculate-branches build step
This was needed when we were building FW8 and FW9 builds from different
branches, but it's no longer useful. The only thing it was still being
used for is in figuring out which branch to tag in the release workflow,
but since we only run that step when the branch is "live" anyway, we can
just hardcode the branch name. Yet more build complexity, begone!