Slash and burn: build complexity begone! (#284)
This PR radically simplifies the build by:

* Getting rid of the split between FW 8 and FW 9 builds now that we no
  longer need to build FW 8 versions
* Getting rid of mono5-sil entirely
* Getting rid of NuGet.targets in favor of a simple dotnet restore
* Getting rid of MsBuild targets in LfMerge.proj in favor of a simple
  dotnet build
* Getting rid of Test target in LfMerge.proj in favor of a simple
  dotnet test
* Getting rid of build/LfMerge.proj entirely and using LfMerge.sln as
  the entry point to the project
* Getting rid of the entire build directory tree as it's no longer needed!

The result is a build that runs much faster, won't fail when TeamCity
goes down, and uses standard dotnet tools to build. With this change,
we'll be able to run local builds on developer machines, without needing
to run them inside a Docker image. That also allows debugging unit tests
in VS Code using standard VS Code debugging tools.

Detailed commit messages from this PR:

* No more FW8 branches

We now build for FW9 only, which mean we no longer need the entire
fieldworks8-* branch system. The calculate-branches and pbuild scripts
can become way simpler.

* Slash and burn: complexity begone!

* Rely entirely on dotnet restore for packages

This lets us get rid of the NuGet.targets file and its complexity.

* Get unit tests running

A lot of failures, but now `dotnet test` is able to find and run the
tests. We'll work on the test failures later.

* Slash and burn: mono5-sil et al, begone!

* Have pbuild.sh create base images directly

Now that mono-sil5 and its complexity are gone, we can have pbuild.sh
create the base images directly since that doesn't take very long. This
will allow us to iterate more quickly on base image changes.

* Slash and burn: LfMerge.proj, begone! Hurrah!

We no longer need to use LfMerge.proj at all! We can now rely entirely
on the standard `dotnet build` and `dotnet test` build process.

* No longer need to keep NuGet packages in build dir

* Slash and burn: build complexity begone!

We can move most things into a single build-and-test.sh script.

* Slash and burn: most of environ script, begone!

We only need a few environment variables (such as adjusting the PATH to
point to the version of Mercurial that we need to use), and soon we'll
be able to get rid of those as well.

* Mount user's home nuget folder as cache

Instead of using a non-standard `/storage/nuget` folder that won't be on
most people's computers, use the NuGet package cache that they alreday
have, located in `~/.nuget/packages`. Because if they're using VS Code
or any other IDE that knows about .NET and NuGet, that folder is
probably already populated with the appropriate packages, which makes
even the first build go faster.

* Slightly simpler handling of SIL.Chorus.Mercurial

PrivateAssets="All" is a simpler way to pull in the build targets we
need from that package.

* Honor TEST_SPEC env var in unit testing

Set TEST_SPEC to a partial (or full) name of a test class or test method
and `dotnet test` will use a "name contains X" filter to locate tests.

* VS Code launch settings for debugging unit tests

To use this, choose the "Run Test Task" command in VS Code. The dotnet
test will run, then pause waiting for a debugger to attach. Go to the
debug menu and choose the "Attach" configuration, then enter the PID
printed on the test console. (There will be two, one for LfMerge.Tests
and one for LfMerge.Core.Tests; most of the tests are in LfMerge.Core so
that's the one you'll need most of the time). The debugger will attach
to the test process, which is still in a paused state. Set any
breakpoints you want to hit in the unit tests, then unpause the debugger
and it will start running the unit tests.

* Stop caching in LfMergeSettings so unit tests work

LfMergeSettings was caching the value of environment variables, but some
strings like BaseDir were being accessed in the constructor, before
derived classes like LfMergeSettingsDouble had a chance to change them.
This meant that many tests, which relied on testlangproj or other
projects being copied into the test's temp dir, were failing because
LfMergeSettings was using the default /var/lib/languageforge/... base
directory instead. By removing the caching, we ensure that the
LfMergeSettingsDouble class is able to change the BaseDir correctly.

* Fix nine more failing unit tests

ProcessingState tests need their expected results updated to include the
new Error property, and the MercurialServer code needs to update its
"shutdown the server" logic to expect "kill" to be a shell builtin
rather than a separate binary. That fixes a total of 9 failing tests.

* Allow testing uncommitted code

By removing the `git checkout` and `git clean` steps, we ensure that we
can run the build process against uncomitted code in the repository.
This will make testing changes much simpler.

A later commit will probably remove the copying of the Git repo into a
temporary working directory, and simply run the build in the user's repo
via a bind mount. That will mean that build outputs (such as TestResult
directories) will end up in your repo, which will be useful in any
situation where you want to debug the unit tests.

* Remove lfmerge.conf (replaced by env vars)

Now that LfMergeSettings uses environment variables, we no longer need
to copy lfmerge.conf into the final Docker image. This resolves a
long-standing TODO.

* Remove rsyslog from final Dockerfile

Now that LfMerge logs to the console rather than syslog, we don't need
rsyslog or its config files in the final Docker image.

* Use same checkout action in entire GHA workflow

One step was still using version 2.4.0 even though the rest of the
workflow had moved to version 3.0.2.

* Remove entire calculate-branches build step

This was needed when we were building FW8 and FW9 builds from different
branches, but it's no longer useful. The only thing it was still being
used for is in figuring out which branch to tag in the release workflow,
but since we only run that step when the branch is "live" anyway, we can
just hardcode the branch name. Yet more build complexity, begone!

* Remove unnecessary step from release workflow

This fulfills a long-standing TODO comment.

* Run build directly in repo (no more copying)

This allows the build to create the installation tarball directly in the
bind-mounted repo, eliminating a copy step.

* Add sanity check to installation script

* Fix workflow syntax error

* Remove previous installation files if they exist

This will avoid the possibility of creating a "mixed" lfmerge
installation containing files from a previous successful build,
alongside a build that only partially succeeded and overwrote some but
not all files from before.

This also remvoes just a little bit of unnecessary chattiness from the
build output.

* Update README with new build instructions

* Remove unused lines from LfMerge.sln

We no longer need the lines about LfMerge.proj or the ones about build
configurations for DbVersions 68, 69, or 70.

* Delete MonoDevelopProperties from LfMerge.sln

We're never going to use MonoDevelop in the future, as VS Code is so
much better. So the whole MonoDevelopProperties section in the .sln file
is a waste of space and processing time. Let's dump the whole thing.

* Remove unneeded test filter in .vscode/tasks.json

Co-authored-by: Christopher Hirt <chris@hirtfamily.net>