Added launch config for debugging in VS Code.
This involved changing the "build" task, which may break other things. This change is a little brittle and thus may not be desirable longterm. I'm certainly open to suggestions. It builds LfMergeQueueManager (specifically) and puts the output in the pbuild.sh output directory. Launching from here uses the full-featured environment already set up for us by the build script.