baseline for working unit tests via pbuild
improvements for finding ChorusMerge, depending on platform
moved LfMerge.Core from .NET Standard to .NET6, to fix warnings about consuming ChorusMerge
this change won't work without a local package of ChorusMerge. it needs the updates to Program.cs. once that PR is merged, we can update the reference here and remove the need for a local nuget source