Using latest LibChorus
Specifying the version in LfMerge.Core, instead of relying on LfMergeBridge to provide one. Once it gets its own update to use a working version, this line can be removed.

This change allows chorusmerge to be found by the LibChorus code, called by FLExBridge. Before, it was pointing to the .exe directly. It was a bug from the .NET upgrade. With this, and the previous commit for FixFwData.exe, LfMerge is now successfully conducting S/R operations.