LfMerge completes S/R's with .NET6 (#306)
* Removing FixFwData and making FixFwData.exe executable

This fixes a bug during sync where the exe couldn't be run due to permissions. I also removed FixFwData, since the production environment was working without it.

* Using latest LibChorus

Specifying the version in LfMerge.Core, instead of relying on LfMergeBridge to provide one. Once it gets its own update to use a working version, this line can be removed.

This change allows chorusmerge to be found by the LibChorus code, called by FLExBridge. Before, it was pointing to the .exe directly. It was a bug from the .NET upgrade. With this, and the previous commit for FixFwData.exe, LfMerge is now successfully conducting S/R operations.