Adding consideration for GUID values in custom lists.
I can't test this, since the List<LfOptionListItem> deserialization just decided to stop working, but I think this is a good general concept.

Assuming LF can't add list items on its own, any time a GUID is stored in MongoDB, it's from LCM. So if the value from MongoDB is a GUID, find the matching value from the LCM list.