Fix "libdl.so not found" error
This is coming from icu.net, which has an icu.net.dll.config file to
redirect the libdl.so requests, but it wasn't copied into the target
directory correctly when it's an indirect dependency. We now explicitly
depend on icu.net so that we can use its dll.config file.