Try uploading files directly, not in tarball
We used to need to upload a tarball because GitHub Actions wasn't
preserving Unix file permissions in the .zip file it would upload, so
things like executable bits were getting stripped. Now that there's an
upload-artifact@v4 that works on different principles internally, let's
see if this step is still necessary. If this procudes Docker images that
still run, we might be able to save some build time.

We also comment out the `dotnet test` step here to save 7 minutes of
waiting for the build; the affected code is after the `dotnet test` so
it will continue to pass but tell us nothing about the changes here.