Temporarily replace flexbridge pkg with project ref
Until the LfMergeBridge changes are merged and a new Nuget package is
produced from them, we'll need to use a ProjectReference so that we can
build LfMerge and test the changes.

The expected location for the flexbridge repo is alongside the LfMerge
repo, in the same parent directory. Also, the correct branch will need
to be checked out in the flexbridge repo.