Don't hardcode GUID in LexBox tests
Rather than assume that the test will be running against the sena-3
project, let's make the test grab an arbitrary entry (the first one that
the LexEntryRepository knows about) and use it. Since the entry whose
GUID we hardcoded was actually the first LexEntry sorted by GUID, this
currently has no effect as we're actually testing the same entry we were
before — but it will allow us to change the test later.