Skip updating project record if there is none
In E2E tests, the SetInputSystems call would fail because there is no
existing project record in the freshly-created, empty Mongo instance.
We don't want to change LfMerge to create the project record, because
the Language Forge code is responsible for doing that. So instead we
simply skip the update if the project record doesn't exist, allowing E2E
tests to pass while not changing the behavior of real code.