End-to-end S/R tests with Lexbox and Mongo (#342)
This end-to-end test runs against a real Lexbox instance, writing data
to a real MongoDB instance rather than a MongoConnectionDouble. It
creates a new project in Lexbox (using the existing sena-3 project as
the initial data), and clones that project into both a LfProject object
and a FwProject object. The FwProject object uses liblcm code to modify
the data, then pushes that modified data to Lexbox. The LfProject then
makes its own changes and goes through the LfMerge Send/Receive process;
because the LfProject's changes are the "ours" in Chorus, the changes
from the LfProject "win" and overwrite the FwProject's changes.

The "real" MongoDB is provided by using a Docker image of Mongo version
6 (via the `mongo:6` tag). This requires a working copy of Docker in the
environment running the tests.

Although there is only one test created here, the SRTestEnvironment
class contains several helper methods that would be useful in many other
end-to-end test scenarios.