Use real MongoDB connection for E2E tests (#348)
A real MongoConnection is used for E2E tests, so that they're truly
end-to-end, rather than the mock one that stores data in memory. The
docker image `mongo:6` is used to provide the MongoDB container.

* Skip updating project record if there is none

In E2E tests, the SetInputSystems call would fail because there is no
existing project record in the freshly-created, empty Mongo instance.
We don't want to change LfMerge to create the project record, because
the Language Forge code is responsible for doing that. So instead we
simply skip the update if the project record doesn't exist, allowing E2E
tests to pass while not changing the behavior of real code.

* Delete MongoDB container after test

Unless LFMERGE_E2E_LEAVE_MONGO_CONTAINER_RUNNING_ON_FAILURE is set (to
any non-empty value except "false" or "0"), we'll tear down the MongoDB
container after tests. Also, if the test succeeds, the MongoDB container
is torn down even if that env var is set. The only time the container is
retained is if the test fails *and* that env var is set.