Fix unit test broken by LfMergeBridge's new logic
LfMergeBridge has become smarter about finding the location of FixFwData
so the test that checks what happens if it's missing now has to work a
lot harder to hide it. It used to be enough to simply change the current
working directory, but now we have to actually hide the FixFwData file
AND the FixFwData.exe file by renaming them temporarily.