Use chg rather than hg for Send/Receive, saving about 20 seconds per S/R (#353)
* Bump Mercurial and Chorus packages to use chg
* Preserve Chorus env vars, ensure chg can find hg
* Delete no-longer-needed test

This test is checking for a failure condition in LfMergeBridge that no
longer happens, and is hard to get running in parallel. We might as well
delete it and keep the code simple.

* Use LfMergeBridge version from before breaking change to Chorus

Later LfMergeBridge NuGet packages were accidentally compiled against a
version of Chorus with a breaking change in the localization manager,
which causes ChorusMerge failures. We'll pin LfMergeBridge to the
version just before the breaking change, just in case.

Once Chorus v7 has a NuGet package, we can upgrade Chorus again.